moved config classes to a seperate folder
