--- src/main/java/de/unistuttgart/overworldbackend/data/LectureConfig.java
@@ -1,13 +1,0 @@
-package de.unistuttgart.overworldbackend.data;
-
-import java.util.List;
-import lombok.AccessLevel;
-import lombok.Data;
-import lombok.experimental.FieldDefaults;
-
-@Data
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class LectureConfig {
-
-  List<WorldConfig> worlds;
-}
