+++ src/main/java/de/unistuttgart/overworldbackend/data/config/DungeonConfig.java
@@ -1,15 +1,13 @@
-package de.unistuttgart.overworldbackend.data;
+package de.unistuttgart.overworldbackend.data.config;
 
-import java.util.List;
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class WorldConfig {
+public class DungeonConfig {
 
-  List<DungeonConfig> dungeons;
   String staticName;
   int numberOfMinigames;
   int numberOfNPCs;
