+++ src/main/java/de/unistuttgart/overworldbackend/data/config/LectureConfig.java
@@ -1,4 +1,6 @@
-package de.unistuttgart.overworldbackend.data;
+package de.unistuttgart.overworldbackend.data.config;
+
+import java.util.List;
 
 import lombok.AccessLevel;
 import lombok.Data;
@@ -6,9 +8,7 @@ import lombok.experimental.FieldDefaults;
 
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class DungeonConfig {
+public class LectureConfig {
 
-  String staticName;
-  int numberOfMinigames;
-  int numberOfNPCs;
+  List<WorldConfig> worlds;
 }
